+++ services/web/server/src/simcore_service_webserver/studies_dispatcher/_controller/rest/redirects.py
@@ -1,52 +1,38 @@
 """Handles request to the viewers redirection entrypoints"""
 
-import functools
 import logging
 
 from aiohttp import web
-from common_library.error_codes import create_error_code
-from common_library.user_messages import user_message
-from models_library.function_services_catalog._utils import ServiceNotFound
 from models_library.projects import ProjectID
 from models_library.projects_nodes_io import NodeID
-from pydantic import ValidationError
-from servicelib.aiohttp import status
 from servicelib.aiohttp.requests_validation import parse_request_query_parameters_as
-from servicelib.aiohttp.typing_extension import Handler
-from servicelib.logging_errors import create_troubleshootting_log_kwargs
 
-from ..dynamic_scheduler import api as dynamic_scheduler_service
-from ..exception_handling import create_error_context_from_request
-from ..products import products_web
-from ..utils import compose_support_error_msg
-from ..utils_aiohttp import create_redirect_to_page_response, get_api_base_url
-from ._catalog import ValidService, validate_requested_service
-from ._constants import MSG_UNEXPECTED_DISPATCH_ERROR
-from ._controller_rest_redirects_schemas import (
+from ....dynamic_scheduler import api as dynamic_scheduler_service
+from ....products import products_web
+from ....utils_aiohttp import create_redirect_to_page_response, get_api_base_url
+from ..._catalog import ValidService, validate_requested_service
+from ..._core import validate_requested_file, validate_requested_viewer
+from ..._errors import (
+    InvalidRedirectionParamsError,
+)
+from ..._models import ServiceInfo, ViewerInfo
+from ..._projects import (
+    get_or_create_project_with_file,
+    get_or_create_project_with_file_and_service,
+    get_or_create_project_with_service,
+)
+from ..._users import UserInfo, ensure_authentication, get_or_create_guest_user
+from ...settings import get_plugin_settings
+from .redirects_exceptions import handle_errors_with_error_page
+from .redirects_schemas import (
     FileQueryParams,
     RedirectionQueryParams,
     ServiceAndFileParams,
     ServiceQueryParams,
 )
-from ._core import validate_requested_file, validate_requested_viewer
-from ._errors import (
-    FileToLargeError,
-    GuestUserNotAllowedError,
-    GuestUsersLimitError,
-    IncompatibleServiceError,
-    InvalidRedirectionParamsError,
-    ProjectWorkbenchMismatchError,
-)
-from ._models import ServiceInfo, ViewerInfo
-from ._projects import (
-    get_or_create_project_with_file,
-    get_or_create_project_with_file_and_service,
-    get_or_create_project_with_service,
-)
-from ._users import UserInfo, ensure_authentication, get_or_create_guest_user
-from .settings import get_plugin_settings
 
 _logger = logging.getLogger(__name__)
+
 
 #
 # HELPERS
@@ -71,18 +57,6 @@ def _create_redirect_response_to_view_page(
     )
 
 
-def _create_redirect_response_to_error_page(
-    app: web.Application, message: str, status_code: int
-) -> web.HTTPFound:
-    # NOTE: these are 'error' page params and need to be interpreted by front-end correctly!
-    return create_redirect_to_page_response(
-        app,
-        page="error",
-        message=message,
-        status_code=status_code,
-    )
-
-
 def _create_service_info_from(service: ValidService) -> ServiceInfo:
     values_map = {
         "key": service.key,
@@ -95,138 +69,12 @@ def _create_service_info_from(service: ValidService) -> ServiceInfo:
     return ServiceInfo.model_construct(_fields_set=set(values_map.keys()), **values_map)
 
 
-def _create_error_redirect_with_logging(
-    request: web.Request,
-    err: Exception,
-    *,
-    message: str,
-    status_code: int,
-    tip: str | None = None,
-) -> web.HTTPFound:
-    """Helper to create error redirect with consistent logging"""
-    error_code = create_error_code(err)
-    user_error_msg = compose_support_error_msg(msg=message, error_code=error_code)
-
-    _logger.exception(
-        **create_troubleshootting_log_kwargs(
-            user_error_msg,
-            error=err,
-            error_code=error_code,
-            error_context=create_error_context_from_request(request),
-            tip=tip,
-        )
-    )
-
-    return _create_redirect_response_to_error_page(
-        request.app,
-        message=user_error_msg,
-        status_code=status_code,
-    )
-
-
-def _create_simple_error_redirect(
-    request: web.Request,
-    public_error: Exception,
-    *,
-    status_code: int,
-) -> web.HTTPFound:
-    """Helper to create simple error redirect without logging
-
-    WARNING: note that the `public_error` is exposed as-is in the user-message
-    """
-    user_error_msg = user_message(
-        f"Unable to open your project: {public_error}", _version=1
-    )
-    return _create_redirect_response_to_error_page(
-        request.app,
-        message=user_error_msg,
-        status_code=status_code,
-    )
-
-
-def _handle_errors_with_error_page(handler: Handler):
-    @functools.wraps(handler)
-    async def wrapper(request: web.Request) -> web.StreamResponse:
-        try:
-            return await handler(request)
-
-        except (web.HTTPRedirection, web.HTTPSuccessful):
-            # NOTE: that response is a redirection that is reraised and not returned
-            raise
-
-        except GuestUserNotAllowedError as err:
-            raise _create_redirect_response_to_error_page(
-                request.app,
-                message=user_message(
-                    "Access is restricted to registered users.<br/><br/>"
-                    "If you don't have an account, please contact support to request one.<br/><br/>",
-                    _version=2,
-                ),
-                status_code=status.HTTP_401_UNAUTHORIZED,
-            ) from err
-
-        except ProjectWorkbenchMismatchError as err:
-            raise _create_error_redirect_with_logging(
-                request,
-                err,
-                message=MSG_UNEXPECTED_DISPATCH_ERROR,
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                tip="project might be corrupted",
-            ) from err
-
-        except (
-            ServiceNotFound,
-            FileToLargeError,
-            IncompatibleServiceError,
-            GuestUsersLimitError,
-        ) as err:
-            raise _create_simple_error_redirect(
-                request,
-                err,
-                status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
-            ) from err
-
-        except (InvalidRedirectionParamsError, web.HTTPUnprocessableEntity) as err:
-            # Validation error in query parameters
-            raise _create_error_redirect_with_logging(
-                request,
-                err,
-                message=user_message(
-                    "The link you provided is invalid because it doesn't contain valid information related to data or a service. "
-                    "Please check the link and make sure it is correct.",
-                    _version=1,
-                ),
-                status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
-                tip="The link might be corrupted",
-            ) from err
-
-        except web.HTTPClientError as err:
-            raise _create_error_redirect_with_logging(
-                request,
-                err,
-                message="Fatal error while redirecting request",
-                status_code=err.status_code,
-                tip="The link might be corrupted",
-            ) from err
-
-        except (ValidationError, web.HTTPServerError, Exception) as err:
-            raise _create_error_redirect_with_logging(
-                request,
-                err,
-                message=MSG_UNEXPECTED_DISPATCH_ERROR,
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                tip="Unexpected failure while dispatching study",
-            ) from err
-
-    return wrapper
-
-
 #
-# API HANDLERS
+# ROUTES
 #
 
 
-@_handle_errors_with_error_page
+@handle_errors_with_error_page
 async def get_redirection_to_viewer(request: web.Request):
     """
     - validate request
