+++ services/web/server/tests/unit/isolated/test_studies_dispatcher_models.py
@@ -13,7 +13,7 @@ from aiohttp.test_utils import make_mocked_request
 from models_library.utils.pydantic_tools_extension import parse_obj_or_none
 from pydantic import ByteSize, TypeAdapter
 from servicelib.aiohttp.requests_validation import parse_request_query_parameters_as
-from simcore_service_webserver.studies_dispatcher._controller_rest_redirects_schemas import (
+from simcore_service_webserver.studies_dispatcher._controller.rest.redirects_schemas import (
     FileQueryParams,
     ServiceAndFileParams,
 )
